Don't build super.img in development tree
We use super.img on FP4 only for release/factory packaging. It's not
useful in development environments, `fastboot flashall` will flash the
logical partitions instead. For release builds we have to generate the
super.img from the target_files anyway.

Meanwhile, also clean up BOARD_DYNAMIC_PARTITION_ENABLE, it's only used
in the Qualcomm makefiles which we don't have in the build anymore.

Change-Id: Ib344f1e220183dc9b6632250a1db4b4b9635fd9d