lights: Drop soong namespace declaration and use ndk backend
This reverts commit 84001c73d63d16a3aaf63598c288965312d9b554.

The ndk_platform backend will soon be deprecated because the ndk backend
can serve the same purpose. This is to eliminate the confusion about
having two variants (ndk and ndk_platform) for the same ndk backend.

Also, the namespace declaration is no more valid since the source
shall be used only for Android 13 and beyond.

Issue: FP4-INT#42
Change-Id: Ie64f8a7a95dafeb4986c8c5aa2303ec3b18f6642